+++ association/lib/readInheritanceFile.py
